Added: Store employee payroll into a file

+++ src/test/java/com/bridgelabz/NIOFileAPITest.java
@@ -12,54 +12,38 @@ import java.util.stream.IntStream;
 public class NIOFileAPITest {
 
     private static String HOME = System.getProperty("user.home");
-    private static String PLAY_WITH_NIO = "TempPlayGround";
+    private static String PLAY_WITH_NOI = "TempPlayGround";
 
     @Test
-    public void givenPathWhenCheckedThenConfirm() throws IOException {
-        /*
-         * Checking if files exists
-         */
+    public void givenPath_WhenChecked_ThenConfirmed()throws IOException {
+        //Check file exists
         Path homePath = Paths.get(HOME);
         Assertions.assertTrue(Files.exists(homePath));
 
-        /*
-         * Delete files and check File not Exist
-         */
-        Path playPath = Paths.get(HOME + "/" + PLAY_WITH_NIO);
-
-        if (Files.exists(playPath))
-            FilesUtils.deleteFiles(playPath.toFile());
+        //Delete File and check File not exists
+        Path playPath = Paths.get(HOME + "/" + PLAY_WITH_NOI);
+        if (Files.exists(playPath)) FilesUtils.deleteFiles(playPath.toFile());
         Assertions.assertTrue(Files.notExists(playPath));
 
-        /*
-         * Create Directory
-         */
-        Files.createDirectory(playPath);
+        //Create Directory
+        Files.createDirectories(playPath);
         Assertions.assertTrue(Files.exists(playPath));
 
-        /*
-         * Create Files
-         */
-        IntStream.range(0, 10).forEach(counter -> {
-            Path tempFile = Paths.get(playPath + "/temp" + counter);
+        //Create File
+        IntStream.range(1, 10).forEach(cntr -> {
+            Path tempFile = Paths.get(playPath + "/temp" + cntr);
             Assertions.assertTrue(Files.notExists(tempFile));
-            /*
-             * Using try catch block for exception handling
-             */
             try {
                 Files.createFile(tempFile);
-
-            } catch (IOException e) {
-                Assertions.assertTrue(Files.exists(tempFile));
+            }catch (IOException e){
+                System.out.println("Error is" + e);
             }
+            Assertions.assertTrue(Files.exists(tempFile));
         });
 
-        /*
-         * List Files, Directories as well as Files with extension
-         */
+        // List Files,Directories as well as Files with Extension
         Files.list(playPath).filter(Files::isRegularFile).forEach(System.out::println);
         Files.newDirectoryStream(playPath).forEach(System.out::println);
-        Files.newDirectoryStream(playPath, path -> path.toFile().isFile() && path.toString().startsWith("temp"))
-                .forEach(System.out::println);
+        Files.newDirectoryStream(playPath, path -> path.toFile().isFile() && path.toString().startsWith("temp")).forEach(System.out::println);
     }
 }
